Implemented correct coordinate usage
To the user a coordinate is expressed as (column, row). This is used
throughout all methods. Internally an array is used, so representation is
different for the array. To solve this helper methods have been introduced
which convert the column and row coordinates into the internal array
equivalent. This is done automatically behind the scenes.