Implemented primitive computer move
Computer now makes a move after a player move. At the moment a random tile
is generated (random type and random rotation). It is then inserted
randomly into either a column or row. The last computer move is recorded
and displayed to the user - at the moment it is considered debug
information but a UI element will be added later.