More major refactoring of UI code
Raw board drawing code is now implemented in the UserInterface. This means
that different interfaces could in theory be used in future (e.g. GUI).
Substantially improved error handling and reporting within the game. The
main game loop now follows the rules for a round correctly - a round
consists of a player tile move, then a player token move, then if the
player hasn't won a computer move. This is now been correctly reflected.