Implemented computer move capabilities
Some changes were made as a result to how the move tile methods work. Both
insertTileRow() and insertTileColumn() now return a Player object. They
also now accept a Player object called performer - this is the player, be
it a computer or human, who is performing the move. The methods then
return an updated version of this to reflect the new spare tile, score,
etc. As a result, the computer is very much like a player in that it
retains a score and spare tile. The computers last move is displayed in
the HUB to make identification easier. Overall, this means rounds are
fully implemented: human tile move; human token move; then computer tile
move.