Removed skip functionality
Player shouldn't be able to skip a token move. This was originally
implemented because they could get stuck on a tile with no where to go.
But it's always possible for them to insert a tile into an adjacenet
position to allow movement. If the user finds themselves stuck, then they
have lost the game. Much like Solitaire, if you fuck yourself up, it's
your fault.