Implemented error handling for Board methods
Specifically, the getTile() and setTile() methods now check if arguments
are within range. If not, an IllegalArgumentException is thrown with
helpful information. This is then caught by the insertTile*() methods in
the event the user has specified an out of range row or column. This
program no longer crashes as a result of human error.