Implemented leaderboard fully
The leaderboard is now displayed at the end. The user is asked whether or
not they would like to submit their score. If they wish to do so, it is
inserted into the leaderboard. The leaderboard is then displayed - by
default only the first 10 entries or less are displayed. The leaderboard
is in ascending order, and the file is assumed to be in ascending order.