Implemented windows and fixed window problems
Maximum length of a line in window content is 72 characters (the full
width of a window is 80 characters, the standard for a terminal). As such,
the help file has been edited to reflect this new change. Empty lines are
now treated correctly (padded out to a space to replace the % symbol). The
main menu and help windows have been fully implemented, whilst the others
remain incomplete.