Code clean up and repackaging/renaming
Labyrinth now exists within the base package as it's the expected main
entry point to the game. UI has been renamed to UserInterface for better
readability. Paths are now treated differently - before it was the case of
using System.getProperty("user.dir"), but now paths are just
"path/to/file" for sake of simplicity. This makes it possible for outside
paths to be used (e.g. /home/john/my_file).