Improved GameView and GameManager substantially
GameView now has a more elegant flow of control when prompting for input.
There is now a help command which gives the user a list of available
commands within the command parser for when they get stuck. The
GameManager has been improved by fixing the token and treasure position
bug. Also, the round, score and current move are now recorded in a saved
game file. Also added a goodbye message in onTreasureChaseWin() which
displays a goodbye message and exits the application (otherwise, the loop
would continue - perhaps running should be used instead, hmm?)