Fixed leaderboard ordering bug
For example, if the leaderboard consisted of 0's and the user got a score
of 1, the score should come after the 0's (0 is a better score than 1).
But that was not the case because the default insertion index was 0 if all
checks failed instead of correctly being inserted in at the end.